+++ test/unit/support/rss-text-parser.ts
@@ -52,7 +52,7 @@ describe('textToHTML function', () => {
       Dare frame thy fearful #symmetry?`;
     const expected = `<p>Tiger, <a href="http://localhost:31337/tiger">@tiger</a>, burning bright<br />
 In the forests of the night,&lt;br&gt;<br />
-What <a href="http://immortal.com/">immortal.com</a> hand or eye<br />
+What <a href="https://immortal.com/">immortal.com</a> hand or eye<br />
 Dare frame thy fearful <a href="http://localhost:31337/search?qs=%23symmetry">#symmetry</a>?</p>`;
     expect(textToHTML(input), 'to be', expected);
   });