+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.18.0] - Not released
+### Changed
+- FreeFeed server now uses Socket.IO v4 for realtime messaging. The previous
+  versions of Socket.IO _clients_ (v2 and v3) are also supported.
 
 ## [2.17.0] - 2023-12-21
 ### Added