+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.18.0] - Not released
+### Changed
+- FreeFeed server now uses Socket.IO v4 for realtime messaging. The previous
+  versions of Socket.IO _clients_ (v2 and v3) are also supported.
+
 ## [2.17.0] - 2023-12-21
 ### Added
 - Users can now subscribe to all comments of posts commented by them. For this